+++ projects/rocprofiler-systems/source/lib/rocprof-sys/library/pmc/collectors/gpu/perfetto_policy.hpp
@@ -446,11 +446,11 @@ private:
         if(effective_metrics.bits.memory_usage && memory_it != tracks.end() &&
            !memory_it->second.track_indexes.empty())
         {
-            const auto usage =
-                rocprofsys::common::units::data_size_cast<
-                    rocprofsys::common::units::megabytes>(
-                    rocprofsys::common::units::bytes{ metric_values.memory_usage })
-                    .count();
+            const auto usage = rocprofsys::common::units::data_size_cast<
+                                   rocprofsys::common::units::megabytes>(
+                                   rocprofsys::common::units::bytes{
+                                       static_cast<double>(metric_values.memory_usage) })
+                                   .count();
             TRACE_COUNTER(
                 "device_memory_usage",
                 counter_track::at(device_index, memory_it->second.track_indexes[0]), ts,
